requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
faiss-cpu==1.10.0
filelock==3.17.0
flake8==7.1.2
flatbuffers==25.2.10
fsspec==2025.2.0
gast==0.6.0
git-filter-repo==2.47.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.29.2
idna==3.10
imbalanced-learn==0.13.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==2.1.3
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-whois==0.9.5
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.1
setuptools==75.8.2
six==1.17.0
sklearn-compat==0.1.3
sqlparse==0.5.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
tf_keras==2.19.0
threadpoolctl==3.5.0
tldextract==5.1.3
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
whois==1.20240129.2
wrapt==1.17.2
